move runner into its own directory
