+++ Tests/MongoSwiftTests/SpecTestRunner/TestOperation.swift
@@ -1,193 +1,4 @@
-import Foundation
 @testable import MongoSwift
-import Nimble
-import XCTest
-
-/// Struct representing the contents of a collection after a spec test has been run.
-internal struct CollectionTestInfo: Decodable {
-    /// An optional name specifying a collection whose documents match the `data` field of this struct.
-    /// If nil, whatever collection used in the test should be used instead.
-    let name: String?
-
-    /// The documents found in the collection.
-    let data: [Document]
-}
-
-/// Struct representing an "outcome" defined in a spec test.
-internal struct TestOutcome: Decodable {
-    /// Whether an error is expected or not.
-    let error: Bool?
-
-    /// The expected result of running the operation associated with this test.
-    let result: TestOperationResult?
-
-    /// The expected state of the collection at the end of the test.
-    let collection: CollectionTestInfo
-}
-
-/// Protocol defining the behavior of an individual spec test.
-protocol SpecTest {
-    var description: String { get }
-    var outcome: TestOutcome { get }
-    var operation: AnyTestOperation { get }
-
-    /// Runs the operation with the given context and performs assertions on the result based upon the expected outcome.
-    func run(client: MongoClient,
-             db: MongoDatabase,
-             collection: MongoCollection<Document>,
-             session: ClientSession) throws
-}
-
-/// Default implementation of a test execution.
-extension SpecTest {
-    internal func run(client: MongoClient,
-                      db: MongoDatabase,
-                      collection: MongoCollection<Document>,
-                      session: ClientSession?) throws {
-        var result: TestOperationResult?
-        var seenError: Error?
-        do {
-            result = try self.operation.op.execute(
-                    client: client,
-                    database: db,
-                    collection: collection,
-                    session: session)
-        } catch {
-            if case let ServerError.bulkWriteError(_, _, bulkResult, _) = error {
-                result = TestOperationResult(from: bulkResult)
-            }
-            seenError = error
-        }
-
-        if self.outcome.error ?? false {
-            expect(seenError).toNot(beNil(), description: self.description)
-        } else {
-            expect(seenError).to(beNil(), description: self.description)
-        }
-
-        if let expectedResult = self.outcome.result {
-            expect(result).toNot(beNil(), description: self.description)
-            expect(result).to(equal(expectedResult), description: self.description)
-        }
-        let verifyColl = db.collection(self.outcome.collection.name ?? collection.name)
-        let foundDocs = try Array(verifyColl.find())
-        expect(foundDocs.count).to(equal(self.outcome.collection.data.count))
-        zip(foundDocs, self.outcome.collection.data).forEach {
-            expect($0).to(sortedEqual($1), description: self.description)
-        }
-    }
-}
-
-/// Protocol for allowing conversion from different result types to `BulkWriteResult`.
-/// This behavior is used to funnel the various CRUD results into the `.bulkWrite` `TestOperationResult` case.
-protocol BulkWriteResultConvertible {
-    var bulkResultValue: BulkWriteResult { get }
-}
-
-extension BulkWriteResult: BulkWriteResultConvertible {
-    internal var bulkResultValue: BulkWriteResult { return self }
-}
-
-extension InsertManyResult: BulkWriteResultConvertible {
-    internal var bulkResultValue: BulkWriteResult {
-        return BulkWriteResult(insertedCount: self.insertedCount, insertedIds: self.insertedIds)
-    }
-}
-
-extension InsertOneResult: BulkWriteResultConvertible {
-    internal var bulkResultValue: BulkWriteResult {
-        return BulkWriteResult(insertedCount: 1, insertedIds: [0: self.insertedId])
-    }
-}
-
-extension UpdateResult: BulkWriteResultConvertible {
-    internal var bulkResultValue: BulkWriteResult {
-        var upsertedIds: [Int: BSONValue]?
-        if let upsertedId = self.upsertedId {
-            upsertedIds = [0: upsertedId]
-        }
-
-        return BulkWriteResult(matchedCount: self.matchedCount,
-                               modifiedCount: self.modifiedCount,
-                               upsertedCount: self.upsertedCount,
-                               upsertedIds: upsertedIds)
-    }
-}
-
-extension DeleteResult: BulkWriteResultConvertible {
-    internal var bulkResultValue: BulkWriteResult {
-        return BulkWriteResult(deletedCount: self.deletedCount)
-    }
-}
-
-/// Enum encapsulating the possible results returned from CRUD operations.
-enum TestOperationResult: Decodable, Equatable {
-    /// Crud operation returns an int (e.g. `count`).
-    case int(Int)
-
-    /// Result of CRUD operations that return an array of `BSONValues` (e.g. `distinct`).
-    case array([BSONValue])
-
-    /// Result of CRUD operations that return a single `Document` (e.g. `findOneAndDelete`).
-    case document(Document)
-
-    /// Result of CRUD operations whose result can be represented by a `BulkWriteResult` (e.g. `InsertOne`).
-    case bulkWrite(BulkWriteResult)
-
-    public init?(from doc: Document?) {
-        guard let doc = doc else {
-            return nil
-        }
-        self = .document(doc)
-    }
-
-    public init?(from result: BulkWriteResultConvertible?) {
-        guard let result = result else {
-            return nil
-        }
-        self = .bulkWrite(result.bulkResultValue)
-    }
-
-    public init(from cursor: MongoCursor<Document>) {
-        self = .array(Array(cursor))
-    }
-
-    public init(from decoder: Decoder) throws {
-        if let insertOneResult = try? InsertOneResult(from: decoder) {
-            self = .bulkWrite(insertOneResult.bulkResultValue)
-        } else if let updateResult = try? UpdateResult(from: decoder), updateResult.upsertedId != nil {
-            self = .bulkWrite(updateResult.bulkResultValue)
-        } else if let bulkWriteResult = try? BulkWriteResult(from: decoder) {
-            self = .bulkWrite(bulkWriteResult)
-        } else if let int = try? Int(from: decoder) {
-            self = .int(int)
-        } else if let array = try? [AnyBSONValue](from: decoder) {
-            self = .array(array.map { $0.value })
-        } else if let doc = try? Document(from: decoder) {
-            self = .document(doc)
-        } else {
-            throw DecodingError.valueNotFound(TestOperationResult.self,
-                                              DecodingError.Context(codingPath: decoder.codingPath,
-                                                                    debugDescription: "couldn't decode outcome")
-            )
-        }
-    }
-
-    internal static func == (lhs: TestOperationResult, rhs: TestOperationResult) -> Bool {
-        switch (lhs, rhs) {
-        case let (.bulkWrite(lhsBw), .bulkWrite(rhsBw)):
-            return lhsBw == rhsBw
-        case let (.int(lhsInt), .int(rhsInt)):
-            return lhsInt == rhsInt
-        case let (.array(lhsArray), .array(rhsArray)):
-            return lhsArray.bsonEquals(rhsArray)
-        case let(.document(lhsDoc), .document(rhsDoc)):
-            return lhsDoc.sortedEquals(rhsDoc)
-        default:
-            return false
-        }
-    }
-}
 
 /// Protocol describing the behavior of a spec test "operation"
 protocol TestOperation: Decodable {
@@ -198,7 +9,7 @@ protocol TestOperation: Decodable {
                  session: ClientSession?) throws -> TestOperationResult?
 }
 
-/// Wrapper around a `TestOperation` allowing it to be decoded from a spec test.
+/// Wrapper around a `TestOperation.swift` allowing it to be decoded from a spec test.
 struct AnyTestOperation: Decodable {
     let op: TestOperation
